requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.0.1

